+++ packages/chain-adapters/src/bitcoin/BitcoinChainAdapter.ts
@@ -1,180 +1,67 @@
-import { AssetNamespace, AssetReference, CAIP2, caip2, caip19 } from '@shapeshiftoss/caip'
+import { AssetNamespace, AssetReference, caip2, caip19 } from '@shapeshiftoss/caip'
 import {
   bip32ToAddressNList,
   BTCOutputAddressType,
   BTCSignTx,
   BTCSignTxInput,
   BTCSignTxOutput,
-  HDWallet,
-  PublicKey,
   supportsBTC
 } from '@shapeshiftoss/hdwallet-core'
-import {
-  BIP44Params,
-  chainAdapters,
-  ChainTypes,
-  NetworkTypes,
-  UtxoAccountType
-} from '@shapeshiftoss/types'
-import { bitcoin } from '@shapeshiftoss/unchained-client'
+import { BIP44Params, chainAdapters, ChainTypes, UtxoAccountType } from '@shapeshiftoss/types'
+import * as unchained from '@shapeshiftoss/unchained-client'
 import coinSelect from 'coinselect'
 import split from 'coinselect/split'
-import WAValidator from 'multicoin-address-validator'
 
 import { ChainAdapter as IChainAdapter } from '../api'
 import { ErrorHandler } from '../error/ErrorHandler'
 import {
   accountTypeToOutputScriptType,
   accountTypeToScriptType,
-  bnOrZero,
-  convertXpubVersion,
   getStatus,
   getType,
   toPath,
   toRootDerivationPath
 } from '../utils'
+import { ChainAdapterArgs, UTXOBaseAdapter } from '../utxo/UTXOBaseAdapter'
 
-export interface ChainAdapterArgs {
-  providers: {
-    http: bitcoin.api.V1Api
-    ws: bitcoin.ws.Client
-  }
-  coinName: string
-  chainId?: CAIP2
-}
-
-export class ChainAdapter implements IChainAdapter<ChainTypes.Bitcoin> {
-  private readonly providers: {
-    http: bitcoin.api.V1Api
-    ws: bitcoin.ws.Client
-  }
-
+export class ChainAdapter
+  extends UTXOBaseAdapter<ChainTypes.Bitcoin>
+  implements IChainAdapter<ChainTypes.Bitcoin>
+{
   public static readonly defaultBIP44Params: BIP44Params = {
     purpose: 84, // segwit native
     coinType: 0,
     accountNumber: 0
   }
 
-  private readonly chainId: CAIP2 = 'bip122:000000000019d6689c085ae165831e93'
-
-  // TODO(0xdef1cafe): constraint this to utxo coins and refactor this to be a UTXOChainAdapter
-  coinName: string
-
   constructor(args: ChainAdapterArgs) {
-    if (args.chainId) {
-      try {
-        const { chain } = caip2.fromCAIP2(args.chainId)
-        if (chain !== ChainTypes.Bitcoin) {
-          throw new Error()
-        }
-        this.chainId = args.chainId
-      } catch (e) {
-        throw new Error(`The ChainID ${args.chainId} is not supported`)
-      }
+    super(args)
+    if (!args.chainId) {
+      throw new Error('chainId required')
     }
-    this.providers = args.providers
+    const { chain, network } = caip2.fromCAIP2(args.chainId)
+    if (chain !== ChainTypes.Bitcoin) {
+      throw new Error('chainId must be a bitcoin chain type')
+    }
+    this.chainId = args.chainId
     this.coinName = args.coinName
+    this.assetId = caip19.toCAIP19({
+      chain,
+      network,
+      assetNamespace: AssetNamespace.Slip44,
+      assetReference: AssetReference.Bitcoin
+    })
   }
 
   getType(): ChainTypes.Bitcoin {
     return ChainTypes.Bitcoin
   }
 
-  getCaip2(): CAIP2 {
-    return this.chainId
-  }
-
-  getChainId(): CAIP2 {
-    return this.chainId
-  }
-
-  private async getPublicKey(
-    wallet: HDWallet,
-    bip44Params: BIP44Params,
-    accountType: UtxoAccountType
-  ): Promise<PublicKey> {
-    const path = toRootDerivationPath(bip44Params)
-    const publicKeys = await wallet.getPublicKeys([
-      {
-        coin: this.coinName,
-        addressNList: bip32ToAddressNList(path),
-        curve: 'secp256k1', // TODO(0xdef1cafe): from constant?
-        scriptType: accountTypeToScriptType[accountType]
-      }
-    ])
-    if (!publicKeys?.[0]) throw new Error("couldn't get public key")
-
-    if (accountType) {
-      return { xpub: convertXpubVersion(publicKeys[0].xpub, accountType) }
-    }
-
-    return publicKeys[0]
-  }
-
-  async getAccount(pubkey: string): Promise<chainAdapters.Account<ChainTypes.Bitcoin>> {
-    if (!pubkey) {
-      return ErrorHandler('BitcoinChainAdapter: pubkey parameter is not defined')
-    }
-
-    try {
-      const caip = await this.getCaip2()
-      const { chain, network } = caip2.fromCAIP2(caip)
-      const { data } = await this.providers.http.getAccount({ pubkey: pubkey })
-
-      const balance = bnOrZero(data.balance).plus(bnOrZero(data.unconfirmedBalance))
-
-      return {
-        balance: balance.toString(),
-        chain: ChainTypes.Bitcoin,
-        caip2: caip,
-        caip19: caip19.toCAIP19({
-          chain,
-          network,
-          assetNamespace: AssetNamespace.Slip44,
-          assetReference: AssetReference.Bitcoin
-        }),
-        chainSpecific: {
-          addresses: data.addresses,
-          nextChangeAddressIndex: data.nextChangeAddressIndex,
-          nextReceiveAddressIndex: data.nextReceiveAddressIndex
-        },
-        pubkey: data.pubkey
-      }
-    } catch (err) {
-      return ErrorHandler(err)
-    }
-  }
-
-  buildBIP44Params(params: Partial<BIP44Params>): BIP44Params {
-    return { ...ChainAdapter.defaultBIP44Params, ...params }
-  }
-
   async getTxHistory(
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     input: chainAdapters.TxHistoryInput
   ): Promise<chainAdapters.TxHistoryResponse<ChainTypes.Bitcoin>> {
-    const { pubkey } = input
-
-    if (!pubkey) return ErrorHandler('pubkey parameter is not defined')
-
-    try {
-      const { data } = await this.providers.http.getTxHistory(input)
-      return {
-        page: data.page,
-        totalPages: data.totalPages,
-        transactions: data.transactions.map((tx) => ({
-          ...tx,
-          chain: ChainTypes.Bitcoin,
-          network: NetworkTypes.MAINNET,
-          symbol: 'BTC',
-          chainSpecific: {
-            opReturnData: ''
-          }
-        })),
-        txs: data.txs
-      }
-    } catch (err) {
-      return ErrorHandler(err)
-    }
+    throw new Error('Method not implemented.')
   }
 
   async buildSendTransaction(tx: chainAdapters.BuildSendTxInput<ChainTypes.Bitcoin>): Promise<{
@@ -207,7 +94,7 @@ export class ChainAdapter implements IChainAdapter<ChainTypes.Bitcoin> {
 
       const account = await this.getAccount(pubkey.xpub)
 
-      type MappedUtxos = Omit<bitcoin.api.Utxo, 'value'> & { value: number }
+      type MappedUtxos = Omit<unchained.bitcoin.Utxo, 'value'> & { value: number }
       const mappedUtxos: MappedUtxos[] = utxos.map((x) => ({ ...x, value: Number(x.value) }))
 
       let coinSelectResult
@@ -278,6 +165,10 @@ export class ChainAdapter implements IChainAdapter<ChainTypes.Bitcoin> {
     }
   }
 
+  buildBIP44Params(params: Partial<BIP44Params>): BIP44Params {
+    return { ...ChainAdapter.defaultBIP44Params, ...params }
+  }
+
   async signTransaction(
     signTxInput: chainAdapters.SignTxInput<chainAdapters.ChainTxType<ChainTypes.Bitcoin>>
   ): Promise<string> {
@@ -293,11 +184,6 @@ export class ChainAdapter implements IChainAdapter<ChainTypes.Bitcoin> {
     } catch (err) {
       return ErrorHandler(err)
     }
-  }
-
-  async broadcastTransaction(hex: string): Promise<string> {
-    const broadcastedTx = await this.providers.http.sendTx({ sendTxBody: { hex } })
-    return broadcastedTx.data
   }
 
   async getFeeData({
@@ -327,7 +213,7 @@ export class ChainAdapter implements IChainAdapter<ChainTypes.Bitcoin> {
       pubkey
     })
 
-    type MappedUtxos = Omit<bitcoin.api.Utxo, 'value'> & { value: number }
+    type MappedUtxos = Omit<unchained.bitcoin.Utxo, 'value'> & { value: number }
     const mappedUtxos: MappedUtxos[] = utxos.map((x) => ({ ...x, value: Number(x.value) }))
 
     let fastFee
@@ -420,15 +306,9 @@ export class ChainAdapter implements IChainAdapter<ChainTypes.Bitcoin> {
     return btcAddress
   }
 
-  async validateAddress(address: string): Promise<chainAdapters.ValidAddressResult> {
-    const isValidAddress = WAValidator.validate(address, this.getType())
-    if (isValidAddress) return { valid: true, result: chainAdapters.ValidAddressResultType.Valid }
-    return { valid: false, result: chainAdapters.ValidAddressResultType.Invalid }
-  }
-
   async subscribeTxs(
     input: chainAdapters.SubscribeTxsInput,
-    onMessage: (msg: chainAdapters.SubscribeTxsMessage<ChainTypes.Bitcoin>) => void,
+    onMessage: (msg: chainAdapters.Transaction<ChainTypes.Bitcoin>) => void,
     onError: (err: chainAdapters.SubscribeError) => void
   ): Promise<void> {
     const {
@@ -445,8 +325,8 @@ export class ChainAdapter implements IChainAdapter<ChainTypes.Bitcoin> {
     await this.providers.ws.subscribeTxs(
       subscriptionId,
       { topic: 'txs', addresses },
-      (msg) => {
-        const transfers = msg.transfers.map<chainAdapters.TxTransfer>((transfer) => ({
+      ({ data: tx }) => {
+        const transfers = tx.transfers.map<chainAdapters.TxTransfer>((transfer) => ({
           caip19: transfer.caip19,
           from: transfer.from,
           to: transfer.to,
@@ -455,18 +335,18 @@ export class ChainAdapter implements IChainAdapter<ChainTypes.Bitcoin> {
         }))
 
         onMessage({
-          address: msg.address,
-          blockHash: msg.blockHash,
-          blockHeight: msg.blockHeight,
-          blockTime: msg.blockTime,
-          caip2: msg.caip2,
+          address: tx.address,
+          blockHash: tx.blockHash,
+          blockHeight: tx.blockHeight,
+          blockTime: tx.blockTime,
+          caip2: tx.caip2,
           chain: ChainTypes.Bitcoin,
-          confirmations: msg.confirmations,
-          fee: msg.fee,
-          status: getStatus(msg.status),
-          tradeDetails: msg.trade,
+          confirmations: tx.confirmations,
+          fee: tx.fee,
+          status: getStatus(tx.status),
+          tradeDetails: tx.trade,
           transfers,
-          txid: msg.txid
+          txid: tx.txid
         })
       },
       (err) => onError({ message: err.message })